// src/motion-provider/types.ts
export interface MotionControllerProps {
  /**
   * @description
   * Allows you to pass options to the useInView hook,
   * which is used to control the animation with it's
   * viewport visibility.
   *
   * @see https://motion.dev/docs/react-use-in-view
   */
  configView?: Omit<UseInViewOptions, "root">;
  /**
   * @description
   * A controllered prop to trigger the animation state in 2 ways:
   * - Start animate
   * - Reverse animate
   *
   * Behaves like a mini version of @param [MotionControllerProps]
   * because it works with only 2 way in one flow:
   *
   * Start —> animateBegin & Reverse —> animateRollback
   *
   * @default undefined
   * @type boolean
   * @example
   *
   * const [trigger, setTrigger] = useState(false);
   *
   * <MotionContainer
   *   onClick={() => setTrigger(prev => !prev)}
   *   elementType="div"
   *   animation={{
   *     mode: ["filterBrightnessFade"],
   *     transition: "springy",
   *     duration: 1,
   *   }}
   *   controller={{
   *     trigger,
   *   }}
   *   className="your-css-goes-here"
   * />
   *   <IfThereIsChildComponent />
   * </MotionContainer>
   *
   */
  trigger?: boolean;
  /**
   * @description
   * Indicates whether the animation should be stopped completely.
   * Not recommended for stand-alone use. Powerful with useAnimation
   * hook when passed as a prop to control the animation flow.
   *
   * @default undefined
   * @type {boolean}
   * @example
   * const { onReverse, control } = useAnimationControl();
   * const { isAnimationStopped, reverse } = useAnimation(control);
   *
   * // might be used in any MP component
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeIn", "filterBlurIn"],
   *     transition: "smooth",
   *   }}
   *   controller={{
   *     isAnimationStopped,
   *     reverse
   *   }}
   *   className="your-css-goes-here"
   * />
   */
  isAnimationStopped?: boolean;
  /**
   * @description
   * Indicates whether the animation should be reversed.
   * Not recommended for stand-alone use. Powerful with useAnimation
   * hook when passed as a prop to control the animation flow.
   *
   * @default undefined
   * @type {boolean}
   * @example
   * const { onReverse, control } = useAnimationControl();
   * const { isAnimationStopped, reverse } = useAnimation(control);
   *
   * // might be used in any MP component
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeIn", "filterBlurIn"],
   *     transition: "smooth",
   *   }}
   *   controller={{
   *     isAnimationStopped,
   *     reverse
   *   }}
   *   className="your-css-goes-here"
   * />
   */
  reverse?: boolean;
}

export interface MotionAnimationProps {
  /**
   * @description
   * Predefined animation mode(s) to be applied. MP provides
   * outrageous numbers(75+) of predefined animation modes for
   * you to choose from without worrying about the compexity which
   * is fixed and always hovering around O(n).
   *
   * @default "opacity"
   * @type {AnimationKeys | AnimationKeys[]}
   * @example
   *
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     // yes, that's actually quite all the thing
   *     // to animate without hesitation :)
   *     mode: "fadeIn",
   *     transition: "smooth",
   *   }}
   *   className="your-css-goes-here"
   * />
   */
  mode: AnimationKeys | AnimationKeys[];
  /**
   * @description
   * Predefined animation transition type to be applied.
   * Find the best transition based on your animation needing.
   *
   * @default "default"
   * @type {TransitionKeys}
   * @example
   *
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: "fadeIn",
   *     // powerful typesafe API's enables developer
   *     // to struggle less to choose the right transition
   *     transition: "smooth",
   *   }}
   *   className="your-css-goes-here"
   * />
   */
  transition: TransitionKeys;
  /**
   * @description
   * Delay of the animation applies right before the animation starts.
   * Creates a standby time and you can think of it as a pause before
   * the animation starts.
   *
   * @type {number} delay of the animation in seconds(s).
   * @default 0
   */
  delay?: number;
  /**
   * @description
   * Duration of the animation in seconds(s).
   * It's recommended to keep it between 0.3-1.5
   * seconds.
   *
   * @type {number} duration of the animation in seconds(s).
   * @default 0.5
   * @minimum 0
   *
   */
  duration?: number;
}
// ...
type GeneralHTMLAttributes = Omit<HTMLAttributes<HTMLElement>, "children">;

export interface MotionContainerProps extends GeneralHTMLAttributes {
  /**
   * @description
   * Defines properties that can be
   * mandatoryly used across MP components. It includes animation modes,
   * transitions, delays, and durations. Basically everything you need to
   * make the web better :)
   *
   * @property {AnimationKeys | AnimationKeys[]} mode - animation mode(s)
   * @property {TransitionKeys} transition - animation transition type
   * @property {number | undefined} [delay] - animation delay
   * @property {number} [duration] - animation duration
   *
   * @example
   *
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeIn", "filterBlurIn"],
   *     transition: "smooth",
   *   }}
   *   className="your-css-goes-here"
   * />
   *
   */
  animation: MotionAnimationProps;
  /**
   * @description
   * The elementType prop allows you to specify the HTML element type
   * that will be used as the root element for the MotionContainer.
   *
   * @default "div"
   * @type {React.ElementType}
   */
  elementType: React.ElementType;
  /**
   * @description
   * The children prop is a React node that will be rendered inside
   * the MotionContainer component. It is optional and can be used to
   * add content to the component.
   *
   * @default undefined
   * @type {React.ReactNode}
   */
  children?: React.ReactNode;
  /**
   * @description
   * This is the central scaffold part that you might see as
   * a god of CAS(Centralized Animation System). It's actually
   * standing on the top of each MP components to control
   * and manage the animation process by managing the flow.
   *
   * Highly recommended to use with both
   * @type {UseAnimationControlProps}
   * and @type {UseAnimationProps}
   *
   * @default undefined
   * @typedef {Object} MotionControllerProps
   * @param {boolean} isAnimationStopped
   * @param {boolean} reverseAnimation
   * @param {Partial<UseInViewOptions>} [configView]
   * @param {boolean} [trigger]
   *
   * @example
   *
   * //full example
   *
   * const { onReverse, control } = useAnimationControl();
   * const { isAnimationStopped, reverse } = useAnimation(control);
   *
   * // trigger the animation freely
   * <button onClick={onReverse}>Reverse</button>
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeUp", "translate3dIn"],
   *     transition: "cubicBounce",
   *     duration: 1,
   *   }}
   *   controller={{
   *     trigger: true, // immediately start animation bypassing configView
   *     configView: { once: true, amount: 0.5 } // determine whether to start animation when it's in viewport
   *     isAnimationStopped, // pass the CAS props
   *     reverse // pass the CAS props
   *   }}
   *   className="your-css-goes-here"
   * />
   *   <IfThereIsChildComponent />
   * </MotionContainer>
   *
   */
  controller?: MotionControllerProps;
}
// ...
export interface AnimationObjProps {
  [key: string]: unknown;
}
export type AnimationModule = {
  initial: AnimationObjProps;
  animate: AnimationObjProps;
};
export interface TransitionConfig {
  duration?: number;
  ease?: EasingDefinition | number[];
  delay?: number;
}